# openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/spaces/{id}" : {
      "get" : {
        "tags" : [ "space-controller" ],
        "operationId" : "get_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SpaceView"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "space-controller" ],
        "operationId" : "update_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateSpaceRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SpaceView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documents/{id}" : {
      "get" : {
        "tags" : [ "document-controller" ],
        "operationId" : "get_7",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "document-controller" ],
        "operationId" : "update_4",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateDocumentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "document-controller" ],
        "operationId" : "softDelete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{id}" : {
      "get" : {
        "tags" : [ "document-controller" ],
        "operationId" : "get_8",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "document-controller" ],
        "operationId" : "update_5",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateDocumentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "document-controller" ],
        "operationId" : "softDelete_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/trash/{trashItemId}/restore" : {
      "post" : {
        "tags" : [ "trash-controller" ],
        "operationId" : "restore",
        "parameters" : [ {
          "name" : "trashItemId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RestoreResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/folders/{id}" : {
      "get" : {
        "tags" : [ "folder-controller" ],
        "operationId" : "get_13",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FolderView"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "folder-controller" ],
        "operationId" : "delete_9",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "folder-controller" ],
        "operationId" : "update_11",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateFolderRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FolderView"
                }
              }
            }
          }
        }
      }
    },
    "/api/folders/{id}" : {
      "get" : {
        "tags" : [ "folder-controller" ],
        "operationId" : "get_14",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FolderView"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "folder-controller" ],
        "operationId" : "delete_10",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "folder-controller" ],
        "operationId" : "update_12",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateFolderRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FolderView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/trash" : {
      "get" : {
        "tags" : [ "trash-controller" ],
        "operationId" : "list_17",
        "parameters" : [ {
          "name" : "resourceType",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 50
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TrashPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trash/{trashItemId}" : {
      "get" : {
        "tags" : [ "trash-controller" ],
        "operationId" : "preview",
        "parameters" : [ {
          "name" : "trashItemId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TrashPreview"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "trash-controller" ],
        "operationId" : "purgeNow",
        "parameters" : [ {
          "name" : "trashItemId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
    "/api/v1/trash/spaces/{spaceId}" : {
      "delete" : {
        "tags" : [ "trash-controller" ],
        "operationId" : "softDeleteSpace",
        "parameters" : [ {
          "name" : "spaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trash/folders/{folderId}" : {
      "delete" : {
        "tags" : [ "trash-controller" ],
        "operationId" : "softDeleteFolder",
        "parameters" : [ {
          "name" : "folderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trash/documents/{documentId}" : {
      "delete" : {
        "tags" : [ "trash-controller" ],
        "operationId" : "softDeleteDocument",
        "parameters" : [ {
          "name" : "documentId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "UpdateSpaceRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "minLength" : 1
          },
          "color" : {
            "type" : "string"
          },
          "externalReference" : {
            "type" : "string"
          },
          "defaultVisibility" : {
            "type" : "string"
          },
          "commentPolicy" : {
            "type" : "string"
          }
        },
        "required" : [ "name" ]
      },
      "SpaceView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "color" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string"
          },
          "externalReference" : {
            "type" : "string"
          },
          "defaultVisibility" : {
            "type" : "string"
          },
          "commentPolicy" : {
            "type" : "string"
          },
          "canManage" : {
            "type" : "boolean"
          },
          "isOwner" : {
            "type" : "boolean"
          },
          "isResponsible" : {
            "type" : "boolean"
          },
          "membership" : {
            "type" : "string"
          }
        }
      },
...
      "DocumentPermissions" : {
        "type" : "object",
        "properties" : {
          "canEdit" : {
            "type" : "boolean"
          },
          "canPublish" : {
            "type" : "boolean"
          },
          "canManageAccess" : {
            "type" : "boolean"
          },
          "canComment" : {
            "type" : "boolean"
          },
          "canManageAttestations" : {
            "type" : "boolean"
          }
        }
      },
      "DocumentResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "spaceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "folderId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : "string"
          },
          "docType" : {
            "type" : "string"
          },
          "body" : {
            "type" : "object",
            "additionalProperties" : { }
          },
          "status" : {
            "type" : "string"
          },
          "currentVersionNo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "reliabilityScore" : {
            "type" : "number"
          },
          "reliabilityComputedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "stale" : {
            "type" : "boolean"
          },
          "contentModifiedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "stalenessThresholdDays" : {
            "type" : "integer",
            "format" : "int32"
          },
          "visibility" : {
            "type" : "string"
          },
          "position" : {
            "type" : "integer",
            "format" : "int32"
          },
          "templateId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "templateVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createdBy" : {
            "$ref" : "#/components/schemas/PersonRef"
          },
          "updatedBy" : {
            "$ref" : "#/components/schemas/PersonRef"
          },
          "owner" : {
            "$ref" : "#/components/schemas/PersonRef"
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TagRef"
            }
          },
          "permissions" : {
            "$ref" : "#/components/schemas/DocumentPermissions"
          }
        }
      },
      "PersonRef" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "displayName" : {
            "type" : "string"
          },
          "initials" : {
            "type" : "string"
          }
        }
      },
      "TagRef" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "color" : {
            "type" : "string"
          },
          "governed" : {
            "type" : "boolean"
          }
        }
      },
...
      "UpdateDocumentRequest" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string",
            "minLength" : 1
          },
          "body" : {
            "type" : "object",
            "additionalProperties" : { }
          },
          "changeSummary" : {
            "type" : "string"
          },
          "docType" : {
            "type" : "string"
          },
          "expectedVersionNo" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "body", "title" ]
      },
...
      "RestoreResult" : {
        "type" : "object",
        "properties" : {
          "documents" : {
            "type" : "integer",
            "format" : "int32"
          },
          "folders" : {
            "type" : "integer",
            "format" : "int32"
          },
          "spaces" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "FolderView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "spaceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "parentFolderId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "position" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createdBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "documentCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "folderCount" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "UpdateFolderRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "position" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "TrashItemView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "resourceType" : {
            "type" : "string"
          },
          "resourceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : "string"
          },
          "snapshot" : {
            "type" : "object",
            "additionalProperties" : { }
          },
          "deletedBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "deletedByName" : {
            "type" : "string"
          },
          "deletedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "purgeAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "TrashPage" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TrashItemView"
            }
          },
          "offset" : {
            "type" : "integer",
            "format" : "int32"
          },
          "limit" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "TrashPreview" : {
        "type" : "object",
        "properties" : {
          "trashItemId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "resourceType" : {
            "type" : "string"
          },
          "resourceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "resource" : {
            "type" : "object",
            "additionalProperties" : { }
          }
        }
      },
...
    }
  }
}